Revert "[Feat] 메인페이지 기능 구현"
