+++ README.md
@@ -8,19 +8,19 @@ ___
 ___
 
 ### 1. [필수] 메인 페이지
-- [x] 디자인 및 화면 구성을 최대한 동일하게 해주세요. (사이즈 및 여백도 최대한 맞춰주세요.) ✨
-- [x] 상품 데이터는 아래 dummy data 를 사용합니다. (더미 데이터는 자유롭게 추가 및 수정 가능)
-- [x] 더미 데이터 : 이미지 링크,  상품 리스트 링크  (←링크 권한 없으면 여기 클릭)
-- [x] RecyclerViewer를 이용해 리스트 화면을 만들어주세요.
-- [x] 상단 툴바를 제거하고 풀스크린 화면으로 세팅해주세요. (상태바(시간/배터리 표시하는 최상단바)는 남기고)
-- [x] 상품 이미지는 모서리를 라운드 처리해주세요.
-- [x] 상품 이름은 최대 두 줄이고, 그래도 넘어가면 뒷 부분에 …으로 처리해주세요.
-- [x] 뒤로가기(BACK)버튼 클릭시 종료하시겠습니까? [확인][취소] 다이얼로그를 띄워주세요. (예시 비디오 참고)
-- [x] 상단 종모양 아이콘을 누르면 Notification을 생성해 주세요. (예시 비디오 참고)
-- [x] 상품 가격은 1000단위로 콤마(,) 처리해주세요.
-- [x] 상품 아이템들 사이에 회색 라인을 추가해서 구분해주세요.
-- [x] 상품 선택시 아래 상품 상세 페이지로 이동합니다.
-- [x] 상품 상세페이지 이동시 intent로 객체를 전달합니다. (Parcelize 사용)
+- [ ] 디자인 및 화면 구성을 최대한 동일하게 해주세요. (사이즈 및 여백도 최대한 맞춰주세요.) ✨
+- [ ] 상품 데이터는 아래 dummy data 를 사용합니다. (더미 데이터는 자유롭게 추가 및 수정 가능)
+- [ ] 더미 데이터 : 이미지 링크,  상품 리스트 링크  (←링크 권한 없으면 여기 클릭)
+- [ ] RecyclerViewer를 이용해 리스트 화면을 만들어주세요.
+- [ ] 상단 툴바를 제거하고 풀스크린 화면으로 세팅해주세요. (상태바(시간/배터리 표시하는 최상단바)는 남기고)
+- [ ] 상품 이미지는 모서리를 라운드 처리해주세요.
+- [ ] 상품 이름은 최대 두 줄이고, 그래도 넘어가면 뒷 부분에 …으로 처리해주세요.
+- [ ] 뒤로가기(BACK)버튼 클릭시 종료하시겠습니까? [확인][취소] 다이얼로그를 띄워주세요. (예시 비디오 참고)
+- [ ] 상단 종모양 아이콘을 누르면 Notification을 생성해 주세요. (예시 비디오 참고)
+- [ ] 상품 가격은 1000단위로 콤마(,) 처리해주세요.
+- [ ] 상품 아이템들 사이에 회색 라인을 추가해서 구분해주세요.
+- [ ] 상품 선택시 아래 상품 상세 페이지로 이동합니다.
+- [ ] 상품 상세페이지 이동시 intent로 객체를 전달합니다. (Parcelize 사용)
 
 
 ### 2. [필수] 디테일 페이지
