--- app/src/main/java/com/example/carrot_market/presentation/DetailActivity.kt
@@ -1,16 +1,0 @@
-package com.example.carrot_market.presentation
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.carrot_market.databinding.ActivityDetailBinding
-
-class DetailActivity : AppCompatActivity() {
-    private val binding: ActivityDetailBinding by lazy {
-        ActivityDetailBinding.inflate(layoutInflater)
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-    }
-}
