--- app/src/main/res/drawable/ic_heart_gray.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M21.633,6.647C21.319,5.919 20.866,5.26 20.299,4.706C19.732,4.151 19.064,3.709 18.33,3.405C17.57,3.09 16.754,2.928 15.931,2.93C14.775,2.93 13.648,3.246 12.668,3.844C12.434,3.987 12.211,4.144 12,4.315C11.789,4.144 11.566,3.987 11.332,3.844C10.352,3.246 9.225,2.93 8.07,2.93C7.238,2.93 6.431,3.089 5.67,3.405C4.934,3.71 4.27,4.148 3.701,4.706C3.134,5.26 2.681,5.919 2.367,6.647C2.041,7.404 1.875,8.208 1.875,9.035C1.875,9.816 2.034,10.629 2.351,11.456C2.616,12.148 2.995,12.865 3.48,13.589C4.249,14.735 5.306,15.931 6.619,17.142C8.794,19.151 10.948,20.538 11.039,20.594L11.594,20.951C11.841,21.108 12.157,21.108 12.403,20.951L12.959,20.594C13.05,20.536 15.202,19.151 17.379,17.142C18.691,15.931 19.748,14.735 20.517,13.589C21.002,12.865 21.384,12.148 21.647,11.456C21.963,10.629 22.123,9.816 22.123,9.035C22.125,8.208 21.959,7.404 21.633,6.647ZM12,19.097C12,19.097 3.656,13.751 3.656,9.035C3.656,6.647 5.632,4.711 8.07,4.711C9.783,4.711 11.269,5.667 12,7.064C12.731,5.667 14.217,4.711 15.931,4.711C18.368,4.711 20.344,6.647 20.344,9.035C20.344,13.751 12,19.097 12,19.097Z"
-      android:fillColor="#8C8C8C"/>
-</vector>
