--- app/src/main/res/drawable/ic_menu.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M2.411,5.6C2.305,5.599 2.2,5.618 2.102,5.658C2.003,5.697 1.914,5.756 1.838,5.83C1.763,5.905 1.703,5.993 1.662,6.091C1.621,6.189 1.6,6.294 1.6,6.4C1.6,6.506 1.621,6.611 1.662,6.709C1.703,6.807 1.763,6.895 1.838,6.97C1.914,7.044 2.003,7.103 2.102,7.142C2.2,7.182 2.305,7.202 2.411,7.2H21.611C21.718,7.202 21.823,7.182 21.921,7.142C22.02,7.103 22.109,7.044 22.184,6.97C22.26,6.895 22.32,6.807 22.361,6.709C22.402,6.611 22.423,6.506 22.423,6.4C22.423,6.294 22.402,6.189 22.361,6.091C22.32,5.993 22.26,5.905 22.184,5.83C22.109,5.756 22.02,5.697 21.921,5.658C21.823,5.618 21.718,5.599 21.611,5.6H2.411ZM2.411,11.2C2.305,11.199 2.2,11.218 2.102,11.258C2.003,11.297 1.914,11.356 1.838,11.43C1.763,11.505 1.703,11.594 1.662,11.691C1.621,11.789 1.6,11.894 1.6,12C1.6,12.106 1.621,12.211 1.662,12.309C1.703,12.407 1.763,12.495 1.838,12.57C1.914,12.644 2.003,12.703 2.102,12.742C2.2,12.782 2.305,12.802 2.411,12.8H21.611C21.718,12.802 21.823,12.782 21.921,12.742C22.02,12.703 22.109,12.644 22.184,12.57C22.26,12.495 22.32,12.407 22.361,12.309C22.402,12.211 22.423,12.106 22.423,12C22.423,11.894 22.402,11.789 22.361,11.691C22.32,11.594 22.26,11.505 22.184,11.43C22.109,11.356 22.02,11.297 21.921,11.258C21.823,11.218 21.718,11.199 21.611,11.2H2.411ZM2.411,16.8C2.305,16.799 2.2,16.818 2.102,16.858C2.003,16.897 1.914,16.956 1.838,17.03C1.763,17.105 1.703,17.194 1.662,17.291C1.621,17.389 1.6,17.494 1.6,17.6C1.6,17.706 1.621,17.811 1.662,17.909C1.703,18.007 1.763,18.095 1.838,18.17C1.914,18.244 2.003,18.303 2.102,18.342C2.2,18.382 2.305,18.402 2.411,18.4H21.611C21.718,18.402 21.823,18.382 21.921,18.342C22.02,18.303 22.109,18.244 22.184,18.17C22.26,18.095 22.32,18.007 22.361,17.909C22.402,17.811 22.423,17.706 22.423,17.6C22.423,17.494 22.402,17.389 22.361,17.291C22.32,17.194 22.26,17.105 22.184,17.03C22.109,16.956 22.02,16.897 21.921,16.858C21.823,16.818 21.718,16.799 21.611,16.8H2.411Z"
-      android:fillColor="#000000"/>
-</vector>
