--- app/src/main/res/drawable/ic_search.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M10.4,2.4C5.991,2.4 2.4,5.991 2.4,10.4C2.4,14.809 5.991,18.4 10.4,18.4C12.317,18.4 14.078,17.719 15.458,16.589L20.234,21.366C20.308,21.442 20.396,21.504 20.494,21.546C20.592,21.588 20.697,21.611 20.803,21.611C20.91,21.613 21.015,21.593 21.114,21.552C21.212,21.512 21.302,21.452 21.377,21.377C21.452,21.302 21.512,21.212 21.552,21.114C21.593,21.015 21.613,20.91 21.612,20.803C21.611,20.697 21.588,20.592 21.546,20.494C21.504,20.396 21.442,20.308 21.366,20.234L16.589,15.458C17.719,14.078 18.4,12.317 18.4,10.4C18.4,5.991 14.809,2.4 10.4,2.4ZM10.4,4C13.944,4 16.8,6.856 16.8,10.4C16.8,13.944 13.944,16.8 10.4,16.8C6.856,16.8 4,13.944 4,10.4C4,6.856 6.856,4 10.4,4Z"
-      android:fillColor="#000000"/>
-</vector>
