+++ app/src/main/res/values/colors.xml
@@ -2,8 +2,4 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="gray">#8C8C8C</color>
-    <color name="cattot">#FF7E36</color>
-    <color name="lightBlue">#4AC1DB</color>
-
 </resources>
