+++ app/src/main/res/values/themes.xml
@@ -3,8 +3,6 @@
     <style name="Base.Theme.Carrot_market" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
-        <item name="windowActionBar">false</item>
-        <item name="windowNoTitle">true</item>
     </style>
 
     <style name="Theme.Carrot_market" parent="Base.Theme.Carrot_market" />
